% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.onAttach}
\alias{.onAttach}
\title{.onAttach}
\usage{
.onAttach(libname, pkgname)
}
\arguments{
\item{libname}{character, name of library}

\item{pkgname}{character, name of package}
}
\value{
package startup message, list of newly loaded packages in printable grid
}
\description{
.onAttach
}
