% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_residual_plots.R
\name{gg_residual_plots}
\alias{gg_residual_plots}
\title{gg_residual_plots}
\usage{
gg_residual_plots(.data, items = c(1:3, 7), new_data = NULL, ...)
}
\arguments{
\item{.data}{an lm() regression object}

\item{items}{a vector of integers from 0 to 7; which of 7 plots to select; default 1:3}

\item{new_data}{data set model was trained on if class is \verb{_lm}; default NULL}

\item{...}{parameters to be passed to \code{patchwork::wrap_plots()}}
}
\value{
a patchwork array of ggplot objects
}
\description{
ggplot versions of common residual plots meant to mimic the \code{plot.lm} and \code{ggfortify::autoplot.lm}
residual plots.
}
\examples{
data(freeny, package="datasets")
fit = lm(y ~ ., data=freeny)
gg_residual_plots(fit)

}
