% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_fevdplot}
\alias{ggvar_fevdplot}
\title{ggvar_fevdplot}
\usage{
ggvar_fevdplot(bvarobject, type = "oir", ...)
}
\arguments{
\item{bvarobject}{a \code{bvarest} object from \code{bvartools}}

\item{type}{character, type of fevd to create, no default, allowed types are
c("oir","gir","sir", "sgir"); only "oir" has sum to unity.}

\item{...}{other parameters for \code{bvartools:fevd.bvar()}; see documentation for details.}
}
\value{
a named list of ggplots, named by response variables of model.
Display with patchwork or just plot one list member.
}
\description{
Forecast Error Variance Decomposition plots for \code{bvartools::bvar} objects
from \link[bvartools:bvartools]{bvartools}; plots all responses or any one
}
\examples{
#' # Load data
library(bvartools)
data("e1")
e1 <- diff(log(e1)) * 100

# Generate model data
model <- bvartools::gen_var(e1, p = 2, deterministic = 2,
                 iterations = 100, burnin = 10)

model <- bvartools::add_priors(model)

# Obtain posterior draws
object <- bvartools::draw_posterior(model)

ggvar_fevdplot(object, type="oir", n.ahead=7)

}
