% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_irf}
\alias{ggvar_irf}
\title{ggvar_irf}
\usage{
ggvar_irf(varirf)
}
\arguments{
\item{varirf}{a \code{varirf} object created using the \code{vars} package}
}
\value{
a ggplot
}
\description{
Plots impulse response functions from a \code{vars::varirf} object
}
\examples{
data("Canada", package="vars")
var.2c <- vars::VAR(Canada, p = 2, type = "const")
IRF = vars::irf(var.2c, impulse = "e", response = "prod")
ggvar_irf(IRF)

}
