% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_plot_stability_break}
\alias{ggvar_plot_stability_break}
\title{ggvar_plot_stability_break}
\usage{
ggvar_plot_stability_break(pi, bv)
}
\arguments{
\item{pi}{breakpoint index; obtain from}

\item{bv}{breakpoint value; use pi to obtain from}
}
\value{
ggplot layers to add to a \code{ggvar_plot_stability()} plot
}
\description{
Adds layers to a \code{ggvar_plot_stability()} plot; a vertical line and a text
notation of the max value, at the max of the stability index.
}
\examples{
data("Canada", package="vars")
var.2c <- vars::VAR(Canada, p = 2, type = "const")
stab = vars::stability(var.2c)
splot = (ggvar_plot_stability(stab))[[1]]
stabprocess = stab$stability[[1]]$process
pindex = which.max(abs(stabprocess))
pvalue = round(stabprocess[pindex], 4)
splot + ggvar_plot_stability_break(pi=pindex, bv=pvalue)

}
