% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_plot_stability_lines}
\alias{ggvar_plot_stability_lines}
\title{ggvar_plot_stability_lines}
\usage{
ggvar_plot_stability_lines(a = 1.2516)
}
\arguments{
\item{a}{limit for upper and lower bounds; default 1.2516, taken from
Gemini discussion of limits for \code{\link[strucchange]{efp}}.}
}
\value{
a ggplot
}
\description{
Internal function that adds layers to a \code{\link{ggvar_plot_stability}} plot;
adds horizontal boundary lines. Called by
\code{\link{ggvar_plot_stability}} optionally.
}
\examples{
data("Canada", package="vars")
var.2c <- vars::VAR(Canada, p = 2, type = "const")
stab = vars::stability(var.2c)
ggvar_plot_stability(stab, breaks=FALSE) # lines=T is default

}
