% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTest.R
\name{GrangerPlot}
\alias{GrangerPlot}
\title{GrangerPlot}
\usage{
GrangerPlot(.data)
}
\arguments{
\item{.data}{a dataframe object created by \code{GrangerTable}.}
}
\value{
a list of two plots for each direction of Granger Causality test
}
\description{
Use this function on a data frame if the data column observations
were taken at the same times
}
\details{
Creates list of plots of Granger test p-values for multiple lags from dataframe
created by \code{GrangerTable}.
}
\examples{
\dontrun{
data(moody, package="mbadtools")
moody \%>\% GrangerTable(ICS, Wgrowth) \%>\% GrangerPlot() \%>\% patchwork::wrap_plots(nrow=2)
}
}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOtyhT7YLcuCOMiy1A?e=LMf1Xa}{Examples},
\code{\link[lmtest]{grangertest}}
}
