% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTest.R
\name{GrangerTable}
\alias{GrangerTable}
\title{GrangerTable}
\usage{
GrangerTable(.data, s1, s2, max.lags = 10)
}
\arguments{
\item{.data}{dataframe containing two numeric columns}

\item{s1}{series 1 numeric column name}

\item{s2}{series 2 numeric column name}

\item{max.lags}{integer, maximum number of lags to consider, default =10.}
}
\value{
data frame with lags and p.values of two granger tests
}
\description{
Computes p.values for Granger cause test of two time series, in both directions. Can
be supplied as input to \code{mbadtools::GrangerPlot()}. Use this if the data is in
a dataframe and you are sure the two series observations are taken at the same times.
}
\examples{
data(moody, package="mbadtools")
moody \%>\% GrangerTable(ICS, Wgrowth)
}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOtyhT7YLcuCOMiy1A?e=LMf1Xa}{Examples},
\code{\link[lmtest]{grangertest}}
}
