% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTestTune.R
\name{GrangerTestPvals}
\alias{GrangerTestPvals}
\title{GrangerTestPvals}
\usage{
GrangerTestPvals(ts1, ts2, max.lags = 10)
}
\arguments{
\item{ts1}{time series of type ts}

\item{ts2}{time series of type ts}

\item{max.lags}{number of lags to consider, default = 10}
}
\value{
data frame with max.lags rows and three columns:
lag, the lag number, p12, the p-value for ts1->ts2, and p21, the reverse.
}
\description{
Use this function for two time series that were taken at the same times.
}
\details{
Creates a data frame of p-values for bi-directional Granger Causality test for
multiple lags, using \code{lmtest::grangertest}.
}
\examples{
data(EuStockMarkets, package="datasets")
# extract two time series
DAX = EuStockMarkets[,"DAX"] 
FTSE = EuStockMarkets[,"FTSE"]
# create the p-values table for the two directions 
GrangerTestPvals(DAX, FTSE)

}
