% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTestTune.R
\name{GrangerTestTune}
\alias{GrangerTestTune}
\title{GrangerTestTune}
\usage{
GrangerTestTune(ts1, ts2, max.lags = 10)
}
\arguments{
\item{ts1}{time series}

\item{ts2}{time series}

\item{max.lags}{maximum number of lags to try, default 10}
}
\value{
a list containing two ggplot objects, showing variation in p-value of Granger causality test in each  directionwith number of lags; can be plotted by new('ggmultiplot') or other means.
}
\description{
Use this function two time series that were taken at the same times.
}
\details{
Run a Granger Causality test in both directions for one time series versus another, over a range of different orders
}
\examples{
\dontrun{
data(EuStockMarkets, package="datasets")
# extract two time series
DA = EuStockMarkets[,"DAX"] 
FT = EuStockMarkets[,"FTSE"]
# plot the two plots with patchwork
GrangerTestTune(DA, FT) \%>\% patchwork::wrap_plots(nrow=2) 
}

}
