% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_add_significance.R
\name{gt_add_significance}
\alias{gt_add_significance}
\title{gt_add_significance}
\usage{
gt_add_significance(.data, level = 0.05, digits = 4)
}
\arguments{
\item{.data}{A tidy or glance tibble resulting from an lm model.}

\item{level}{The significance level selected. default .05}
}
\value{
a gt table with the pvalue column colored where result is significant.
}
\description{
Makes a gt table, and adds colored significance notation, to a data frame containing p_value column. 
Useful for glance and tidy output from lm models.
}
\examples{
require(tidyverse);
fit = lm(dist~speed, data=cars);
fit \%>\% broom::tidy() \%>\% add_significance();

}
