% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbadtools-package.R
\docType{package}
\name{mbadtools-package}
\alias{mbadtools}
\alias{mbadtools-package}
\title{mbadtools: Tools for ...}
\description{
Description of the package.
}
\details{
See the package README:
\url{https://github.com/bhartman2/mbadtools#readme}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bhartman2/mbadtools}
  \item Report bugs at \url{https://github.com/bhartman2/mbadtools/Issues}
}

}
\author{
\strong{Maintainer}: Bruce Hartman \email{bruce@ahartman.net}

Authors:
\itemize{
  \item Bruce Hartman \email{bruce@ahartman.net}
}

}
\keyword{internal}
