% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moody.R
\docType{data}
\name{moody}
\alias{moody}
\title{moody}
\format{
A data frame with 299 rows and 3 columns:
\describe{
\item{Date}{date-time of observation}
\item{Wgrowth}{numeric, wage growth at that time}
\item{ICS}{numeric consumer sentiment survey at that time }
}
}
\source{
consumer sentiment data from University of Michigan Survey;
Wage growth from Atlanta Fed Wage Tracker.
}
\usage{
moody
}
\description{
A dataset containing wage growth and consumer sentiment data.
}
\keyword{datasets}
