% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myresids.R
\docType{data}
\name{myresids}
\alias{myresids}
\title{myresids}
\format{
A character vector with 4 entries:
\describe{
\item{resid}{character, residual-fitted plot }
\item{qq}{character, qq plot}
\item{ls}{character, location-scale plot }
\item{hist}{character, histogram of residuals }
}
}
\source{
ggResidpanel package documentation.
}
\usage{
myresids
}
\description{
A character vector containing the four ggResidpanel::resid_panel() plots mbadtools recommends.
}
\keyword{datasets}
