% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_arrange_data}
\alias{pareto_arrange_data}
\title{pareto_arrange_data}
\usage{
pareto_arrange_data(.data, countcol)
}
\arguments{
\item{.data}{data frame with a category variable and a count column,
often produced with \code{dplyr::summarize}}

\item{countcol}{name of the column with the counts, no quotes required}
}
\value{
a dataframe with the category and count columns plus the Cumulative and Percentage columns.
}
\description{
Arrange compute data for a pareto plot
}
\examples{
\dontrun{
library(qcc)
data(mtcars)
A = mtcars \%>\% mutate(cyl= as.character(cyl))
A1 = A \%>\% group_by(cyl) \%>\% summarize(c_count=n())
A2 = pareto_arrange_data(A1, c_count)
print(A2)
}
}
