% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_cutoff_line}
\alias{pareto_cutoff_line}
\title{pareto_cutoff_line}
\usage{
pareto_cutoff_line(maxcount, percent = 80)
}
\arguments{
\item{maxcount}{maximum count}

\item{percent}{cutoff percent, default=80}
}
\value{
added ggplot layer with the Pareto percentage line
}
\description{
Puts a cutoff line on a pareto_plot
}
\examples{
\dontrun{
library(qcc)
library(dplyr)
data(mtcars, package="qcc")
A = mtcars \%>\% mutate(cyl = as.character(cyl))
A1 = A \%>\% group_by(cyl) \%>\% summarize(c_count=n())
A2 = pareto_arrange_data(A1, c_count)
mc = max(A2$c_count)
A2 \%>\% pareto_plot(cyl, c_count) +
  pareto_sec_axis(mc) +
  pareto_pct_labels(mc, nudge_y=1, size=3) +
  pareto_cutoff_line(mc)
  }
}
