% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_pct_labels}
\alias{pareto_pct_labels}
\title{pareto_pct_labels}
\usage{
pareto_pct_labels(maxcount, ...)
}
\arguments{
\item{maxcount}{maximum count}

\item{...}{additional parameters to \code{geom_text}}
}
\value{
added ggplot layer
}
\description{
Add percent labels on cumulative percent line. Use additional parameters
to size and position the labels on the plot. Data object must have a Percentage column.
}
\examples{
\dontrun{
library(qcc)
library(dplyr)
data(mtcars)
A = mtcars \%>\% mutate(cyl = as.character(cyl))
A1 = A \%>\% group_by(cyl) \%>\% summarize(c_count=n())
A2 = pareto_arrange_data(A1, c_count)
mc = max(A2$c_count)
A2 \%>\% pareto_plot(cyl, c_count) +
  pareto_sec_axis(mc) +
  pareto_pct_labels(mc, nudge_y=1, size=3)
}
  
}
