% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_plot}
\alias{pareto_plot}
\title{pareto_plot}
\usage{
pareto_plot(.data, category, count)
}
\arguments{
\item{.data}{a pareto_arranged data frame}

\item{category}{name of the category column}

\item{count}{name of the count column}
}
\value{
a ggplot object
}
\description{
Make a pareto plot from data arranged with \code{pareto_arrange}
}
\examples{
data(chickwts, package="datasets")
K = chickwts \%>\% group_by(feed) \%>\% summarize(avwt = mean(weight))
K1 = pareto_arrange_data(K, avwt)
pareto_plot(K1, feed, avwt)

}
