% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_sec_axis}
\alias{pareto_sec_axis}
\title{pareto_sec_axis}
\usage{
pareto_sec_axis(maxcount)
}
\arguments{
\item{maxcount}{numeric, maximum value of the count column}
}
\value{
a ggplot layer
}
\description{
pareto_sec_axis
}
\examples{
\dontrun{
library(qcc)
library(dplyr)
data(mtcars)
A = mtcars \%>\% mutate(cyl = as.character(cyl))
A1 = A \%>\% group_by(cyl) \%>\% summarize(c_count=n())
A2 = pareto_arrange_data(A1, c_count)
A2 \%>\% pareto_plot(cyl, c_count) +
  pareto_sec_axis(max(A2$c_count))
  }
}
