% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{bvarirf_to_varirf}
\alias{bvarirf_to_varirf}
\title{bvarirf_to_varirf}
\usage{
bvarirf_to_varirf(
  bvarirf,
  impulse,
  response,
  cumulative = TRUE,
  ortho = TRUE,
  boot = TRUE,
  ci = 0.05,
  runs = 100
)
}
\arguments{
\item{bvarirf}{a \code{bvarirf} object made by \code{\link{bvartools}}}

\item{impulse}{character, the impulse column name}

\item{response}{character, the response column name}

\item{cumulative}{logical, default TRUE}

\item{ortho}{logical, default TRUE}

\item{boot}{logical, default TRUE}

\item{ci}{numeric p-value for confidence interval; in (0,1), default .05}

\item{runs}{integer, number of runs, default 100}
}
\value{
a \code{varirf} skeleton object good enough for \code{\link{ggvar_irf}} to plot.
}
\description{
Transforms a \code{bvarirf} object from\code{\link[bvartools]{irf}} into
a \code{varirf} skeleton object from\code{\link[vars]{irf}},
for \code{\link{ggvar_irf}} to plot. The \code{bvarirf} object only contains one response at a time.
}
\examples{
# Load data
data("e1", package="bvartools")
e1 <- diff(log(e1)) * 100

# Generate model data
model <- bvartools::gen_var(e1, p = 2, deterministic = 2,
                 iterations = 100, burnin = 10)
# Number of iterations and burnin should be much higher.

# Add prior specifications
model <- bvartools::add_priors(model)

# Obtain posterior draws
object <- bvartools::draw_posterior(model)

# Calculate IR
ir <- bvartools::irf(object, impulse = "invest", response = "cons")

# Plot IR
X = bvarirf_to_varirf(ir, impulse="invest",  response="cons")
ggvar_irf(X)

}
