% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{bvarirf_to_varirf}
\alias{bvarirf_to_varirf}
\title{bvarirf_to_varirf}
\usage{
bvarirf_to_varirf(
  bvarirf,
  impulse,
  response,
  cumulative = TRUE,
  ortho = TRUE,
  boot = TRUE,
  ci = 0.05,
  runs = 100
)
}
\arguments{
\item{bvarirf}{a \code{bvarirf} object made by \code{bvartools}}

\item{impulse}{character, the impulse column name}

\item{response}{character, the response column name}

\item{cumulative}{logical, default TRUE}

\item{ortho}{logical, default TRUE}

\item{boot}{logical, default TRUE}
}
\value{
a \code{varirf} skeleton object good enough for \code{mbadtools::ggvar_irf()} to plot.
}
\description{
Transforms a \code{bvartools::bvarirf} object into a \code{vars::varirf} skeleton object
for \code{ggvar_irf()} to plot. \code{bvartools} only allows one response at a time
in its \code{bvarirf} object.
}
\examples{
# Load data
data("e1", package="bvartools")
e1 <- diff(log(e1)) * 100

# Generate model data
model <- bvartools::gen_var(e1, p = 2, deterministic = 2,
                 iterations = 100, burnin = 10)
# Number of iterations and burnin should be much higher.

# Add prior specifications
model <- bvartools::add_priors(model)

# Obtain posterior draws
object <- bvartools::draw_posterior(model)

# Calculate IR
ir <- bvartools::irf(object, impulse = "invest", response = "cons")

# Plot IR
X = bvarirf_to_varirf(ir, impulse="invest",  response="cons")
ggvar_irf(X)

}
