% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_partial_residual_plots.R
\name{gg_partial_residual_plots}
\alias{gg_partial_residual_plots}
\title{Partial Residual Plots}
\usage{
gg_partial_residual_plots(.data)
}
\arguments{
\item{.data}{an lm object}
}
\value{
a patchwork plot layout containing the partial residual plots.
}
\description{
Draw partial residual plots with ggplot methods. Automatically plots all driver
partial plots in the model.
}
\examples{
data(freeny, package="datasets")
fit = lm(y ~ ., data = freeny)
gg_residual_plots(fit)
gg_partial_residual_plots(fit)

}
