% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_residual_plots.R
\name{gg_residual_plots}
\alias{gg_residual_plots}
\title{gg_residual_plots}
\usage{
gg_residual_plots(.data, items = 1:3)
}
\arguments{
\item{.data}{an lm() regression object}

\item{items}{a vector of integers from 0 to 7; which of 7 plots to select; default 1:3}
}
\value{
a patchwork array of ggplot objects
}
\description{
ggplot versions of common residual plots meant to mimic the plot.lm() and ggfortify::autoplot.lm()
residual plots.
}
\examples{
data(freeny, package="datasets")
fit = lm(y ~ ., data=freeny)
gg_residual_plots(fit)
gg_residual_plots(fit, items=4:6)

}
