% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_forecastplot}
\alias{ggvar_forecastplot}
\title{ggvar_forecastplot}
\usage{
ggvar_forecastplot(bvar_pred)
}
\arguments{
\item{bvar_pred}{predictions of class \code{bvarprd} or \code{varprd}}
}
\value{
a list of ggplots
}
\description{
plot multiple timeseries and forecasts made with \code{predict()};
use design and \code{patchwork::wrap_plots()} to plot from list
}
\examples{
# Load data
library(bvartools)
data("e1")
e1 <- diff(log(e1)) * 100

# Generate model data
model <- gen_var(e1, p = 2, deterministic = 2,
                 iterations = 100, burnin = 10)

model <- add_priors(model)

# Obtain posterior draws
object <- draw_posterior(model)

# Calculate forecasts
pred <- predict(object, new_d = rep(1, 10))

# Plot forecasts
ggvar_forecastplot(pred)

}
