% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggvar_irf.R
\name{ggvar_plot_stability}
\alias{ggvar_plot_stability}
\title{ggvar_plot_stability}
\usage{
ggvar_plot_stability(varstabil)
}
\arguments{
\item{varstabil}{a \code{varstabil} object created from \code{vars::stability()}}
}
\value{
a list of ggplots one for each equation
}
\description{
Plots the empirical fluctuation processes created by the \code{vars::stability()} test;
one is created for each equation. Use \code{patchwork::wrap_plots(nrow=n)} to display,
where n is the number of equations.
}
\examples{
data("Canada", package="vars")
var.2c <- vars::VAR(Canada, p = 2, type = "const")
stab = vars::stability(var.2c)
ggvar_plot_stability(stab)

}
