% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTest.R, R/GrangerTestTune.R
\name{GrangerPlot}
\alias{GrangerPlot}
\title{GrangerPlot}
\usage{
GrangerPlot(.data)

GrangerPlot(.data)
}
\arguments{
\item{.data}{a dataframe created by \code{GrangerTestPvals}.}
}
\value{
a list of two plots for each direction of Granger Causality test

a list of two plots for each direction of Granger Causality test
}
\description{
Creates list of plots of Granger test p-values for multiple lags from dataframe
created by \code{GrangerTestPvals}.

Creates list of plots of Granger test p-values for multiple lags from dataframe
created by \code{GrangerTestPvals}.
}
\examples{
data(moody, package="datasets")
gpv = GrangerTestPvals(moody, ICS, Wgrowth, max.lags=10)
GrangerPlot(gpv)
data(moody, packge="datasets")
gpv = moody \%>\% GrangerTestPvals(ICS, Wgrowth, max.lags=10)
GrangerPlot(gpv)
}
