% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTest.R
\name{GrangerPlot}
\alias{GrangerPlot}
\title{GrangerPlot}
\usage{
GrangerPlot(.data)
}
\arguments{
\item{.data}{a dataframe object created by \code{GrangerTestPvals}.}
}
\value{
a list of two plots for each direction of Granger Causality test
}
\description{
Creates list of plots of Granger test p-values for multiple lags from dataframe
created by \code{GrangerTestPvals}.
}
\examples{
data(moody, package="datasets")
gpv = GrangerTestPvals(moody, ICS, Wgrowth, max.lags=10)
GrangerPlot(gpv)
}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOtyhT7YLcuCOMiy1A?e=LMf1Xa}{Examples},
\code{\link[lmtest]{grangertest}}
}
