% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTestTune.R
\name{GrangerTestPvals}
\alias{GrangerTestPvals}
\title{GrangerTestPvals}
\usage{
GrangerTestPvals(ts1, ts2, max.lags = 10)
}
\arguments{
\item{ts1}{time series}

\item{ts2}{time series}

\item{max.lags}{integer, maximum number of lags to consider, default =10.}
}
\value{
data frame
}
\description{
Computes p.values for Granger cause test of two time series, in both directions.
}
\examples{
\dontrun{
GrangerTestPvals(moody$ICS, moody$Wgrowth, max.lags=10)
}
}
