% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrangerTest.R
\name{GrangerTestPvals}
\alias{GrangerTestPvals}
\title{GrangerTestPvals}
\usage{
GrangerTestPvals(.data, s1, s2, max.lags = 10)
}
\arguments{
\item{.data}{dataframe containing two series columns}

\item{s1}{series 1 name}

\item{s2}{series 2 name}

\item{max.lags}{integer, maximum number of lags to consider, default =10.}
}
\value{
data frame with lags and p.values of two granger tests
}
\description{
Computes p.values for Granger cause test of two time series, in both directions.
}
\examples{
data(moody, package="mbadtools")
GrangerTestPvals(moody, ICS, Wgrowth)
}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOtyhT7YLcuCOMiy1A?e=LMf1Xa}{Examples},
\code{\link[lmtest]{grangertest}}
}
