% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbadtools-package.R
\docType{package}
\name{mbadtools-package}
\alias{mbadtools}
\alias{mbadtools-package}
\title{mbadtools: MBAD Data Analytics Functions}
\description{
Some functions useful for my MBAD courses on Data Analytics. Loads packages from core packages list on library call Includes Residual plots, Pareto plots, and adding significance color to pvalues in a gt table. Includes Granger causality table and plot functions. Includes functions for handling VAR models with bvartools and plotting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bhartman2/mbadtools}
  \item Report bugs at \url{https://github.com/bhartman2/mbadtools/Issues}
}

}
\author{
\strong{Maintainer}: Bruce Hartman \email{bruce@ahartman.net}

Authors:
\itemize{
  \item Bruce Hartman \email{bruce@ahartman.net}
}

}
\keyword{internal}
