% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbadtools-package.R
\docType{package}
\name{mbadtools-package}
\alias{mbadtools}
\alias{mbadtools-package}
\title{mbadtools: MBAD Data Analytics Functions}
\description{
Some functions useful for my MBAD courses on Data Analytics.
}
\author{
\strong{Maintainer}: Bruce Hartman \email{bruce@ahartman.net}

}
\keyword{internal}
