% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_arrange_data}
\alias{pareto_arrange_data}
\title{pareto_arrange_data}
\usage{
pareto_arrange_data(.data, countcol)
}
\arguments{
\item{.data}{data frame with a category variable and a count column,
often produced with \code{dplyr::summarize}}

\item{countcol}{name of the column with the counts, no quotes required}
}
\value{
a dataframe with the category and count columns plus the Cumulative and Percentage columns.
}
\description{
Arrange compute data for a pareto plot
}
\examples{
data(chickwts, package="datasets")
K = chickwts \%>\% group_by(feed) \%>\% summarize(avwt = mean(weight))
K1 = pareto_arrange_data(K, avwt)
print(K1)

}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOt0X0A8jVrshpaYQg?e=01d2ZT}{Pareto Plot Examples},
\code{\link[qcc]{pareto.chart}}
}
