% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_cutoff_line}
\alias{pareto_cutoff_line}
\title{pareto_cutoff_line}
\usage{
pareto_cutoff_line(maxcount, percent = 80)
}
\arguments{
\item{maxcount}{maximum count}

\item{percent}{cutoff percent, default=80}
}
\value{
added ggplot layer with the Pareto percentage line
}
\description{
Puts a cutoff line on a pareto_plot
}
\examples{
data(chickwts, package="datasets")
K = chickwts \%>\% group_by(feed) \%>\% summarize(avwt = mean(weight))
K1 = pareto_arrange_data(K, avwt)
pareto_plot(K1, feed, avwt) +
  pareto_cutoff_line(max(K1$avwt))

}
\seealso{
\href{https://1drv.ms/u/s!AvpG0vuEDBBQoOt0X0A8jVrshpaYQg?e=01d2ZT}{Pareto Plot Examples},
\code{\link[qcc]{pareto.chart}}
}
