% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_pct_labels}
\alias{pareto_pct_labels}
\title{pareto_pct_labels}
\usage{
pareto_pct_labels(maxcount, ...)
}
\arguments{
\item{maxcount}{maximum count}

\item{...}{additional parameters to \code{geom_text}}
}
\value{
added ggplot layer
}
\description{
Add percent labels on cumulative percent line. Use additional parameters
to size and position the labels on the plot. Data object must have a Percentage column.
}
\examples{
data(chickwts, package="datasets")
K = chickwts \%>\% group_by(feed) \%>\% summarize(avwt = mean(weight))
K1 = pareto_arrange_data(K, avwt)
pareto_plot(K1, feed, avwt) +
  pareto_pct_labels(max(K1$avwt), nudge_y=1, size=3)
  
}
