% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_plot}
\alias{pareto_plot}
\title{pareto_plot}
\usage{
pareto_plot(.data, category, count)
}
\arguments{
\item{.data}{a pareto_arranged data frame}

\item{category}{name of the category column}

\item{count}{name of the count column}
}
\value{
a ggplot object
}
\description{
Make a pareto plot from data arranged with \code{pareto_arrange}
}
\examples{
\dontrun{
library(qcc)
library(dplyr)
data(mtcars)
A = mtcars \%>\% mutate(cyl = as.character(cyl))
A1 = A \%>\% group_by(cyl) \%>\% summarize(c_count=n())
A2 = pareto_arrange_data(A1, c_count)
A2 \%>\% pareto_plot(cyl, c_count)
}
}
