% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto_plot.R
\name{pareto_sec_axis}
\alias{pareto_sec_axis}
\title{pareto_sec_axis}
\usage{
pareto_sec_axis(maxcount)
}
\arguments{
\item{maxcount}{numeric, maximum value of the count column}
}
\value{
a ggplot layer
}
\description{
pareto_sec_axis
}
\examples{
data(chickwts, package="datasets")
K = chickwts \%>\% group_by(feed) \%>\% summarize(avwt = mean(weight))
K1 = pareto_arrange_data(K, avwt)
pareto_plot(K1, feed, avwt) +
  pareto_sec_axis(max(K1$avwt))

}
